Cloud integration tests (MinIO/SeaweedFS) + clippy clean + cleanup
- Rewrite cloud_integration.rs against the new trait API: upload a >16MiB
  capture to a real S3-compatible store and assert byte-range partitioned
  reads (independent range GETs) equal the sequential read, and that a SQL
  query is identical at 1 vs N partitions with exactly one parse pass.
  Gated by the s3 feature + PCAPSQL_S3_TEST_ENDPOINT so the normal test run
  needs no object store. Verified in-session against SeaweedFS.
- testdata/cloud: MinIO (Chainguard) docker-compose with bucket bootstrap,
  SeaweedFS s3.json, and a setup.sh supporting both backends.
- Fix all clippy findings across the workspace; clippy --all-targets
  -D warnings is clean (default and s3 feature).
- Rename the repurposed --stats helper to print_parse_stats; refresh crate
  docs (parallel parsing instead of the removed LRU cache).

+++ pcapsql-core/tests/partition_equivalence.rs
@@ -142,7 +142,7 @@ fn legacy_all_four_variants() {
         let packets: Vec<GenPacket> = (0..20)
             .map(|i| GenPacket {
                 ts_sec: 1_700_000_000 + i,
-                ts_frac: (i as u32 * 12345) % 1_000_000_000,
+                ts_frac: (i * 12345) % 1_000_000_000,
                 data: vec![(i % 256) as u8; 14 + (i as usize % 50)],
                 origlen: 14 + (i % 50),
             })
